+++ src/enums/serviceAliases.ts
@@ -1,9 +1,13 @@
 export default {
-  routeTable: 'routeTables',
-  securityGroup: 'securityGroups',
+  camPolicy: 'camPolicies',
   ccn: 'ccns',
   ccnAttachment: 'ccnAttachments',
+  customerGateway: 'customerGateways',
+  routeTable: 'routeTables',
+  securityGroup: 'securityGroups',
   subnet: 'subnets',
   vpc: 'vpcInstances',
+  vpnConnection: 'vpnConnections',
   vpnGateway: 'vpnGateways',
+  vpnGatewayRoute: 'vpnGatewayRoutes',
 }