+++ src/enums/services.ts
@@ -1,12 +1,16 @@
 export default {
+  camPolicy: 'camPolicy',
   camUser: 'camUser',
+  ccn: 'ccn',
+  ccnAttachment: 'ccnAttachment',
   customerGateway: 'customerGateway',
+  networkAcl: 'networkAcl',
   routeTable: 'routeTable',
   securityGroup: 'securityGroup',
   securityGroupRule: 'securityGroupRule',
-  ccn: 'ccn',
-  ccnAttachment: 'ccnAttachment',
   subnet: 'subnet',
   vpc: 'vpc',
+  vpnConnection: 'vpnConnection',
   vpnGateway: 'vpnGateway',
+  vpnGatewayRoute: 'vpnGatewayRoute',
 }